+++ src/algorithm/segmentation/objectModels/ObjectModelCircle.h
@@ -26,17 +26,17 @@ public:
 	virtual ~ObjectModelCircle(){};
 
 	void getSamples (int &iterations, std::vector<int> &samples);
-	bool computeModelCoefficients (const std::vector<int> &samples, Eigen::VectorXf &model_coefficients);
-	void optimizeModelCoefficients (const std::vector<int> &inliers, const Eigen::VectorXf &model_coefficients,
-			Eigen::VectorXf &optimized_coefficients);
-	void getDistancesToModel (const Eigen::VectorXf &model_coefficients, std::vector<double> &distances);
-	void selectWithinDistance (const Eigen::VectorXf &model_coefficients, double threshold,
+	bool computeModelCoefficients (const std::vector<int> &samples, Eigen::VectorXd &model_coefficients);
+	void optimizeModelCoefficients (const std::vector<int> &inliers, const Eigen::VectorXd &model_coefficients,
+			Eigen::VectorXd &optimized_coefficients);
+	void getDistancesToModel (const Eigen::VectorXd &model_coefficients, std::vector<double> &distances);
+	void selectWithinDistance (const Eigen::VectorXd &model_coefficients, double threshold,
 			std::vector<int> &inliers);
-	void getInlierDistance (std::vector<int> &inliers, const Eigen::VectorXf &model_coefficients,  std::vector<double> &distances);
-	void projectPoints (const std::vector<int> &inliers, const Eigen::VectorXf &model_coefficients,
+	void getInlierDistance (std::vector<int> &inliers, const Eigen::VectorXd &model_coefficients,  std::vector<double> &distances);
+	void projectPoints (const std::vector<int> &inliers, const Eigen::VectorXd &model_coefficients,
 			PointCloud3D* projectedPointCloud);
-	bool doSamplesVerifyModel (const std::set<int> &indices, const Eigen::VectorXf &model_coefficients, double threshold);
-	void computeRandomModel (int &iterations, Eigen::VectorXf &model_coefficients, bool &isDegenerate, bool &modelFound);
+	bool doSamplesVerifyModel (const std::set<int> &indices, const Eigen::VectorXd &model_coefficients, double threshold);
+	void computeRandomModel (int &iterations, Eigen::VectorXd &model_coefficients, bool &isDegenerate, bool &modelFound);
 
 	inline int getNumberOfSamplesRequired(){return 2;};
 };
